coding

+++ src/deploy/tasks/git.js
@@ -13,17 +13,27 @@ module.exports = async (config) => {
   const remotes = response.stdout ? response.stdout.split(/\n/) : [];
   const method = remotes.indexOf("origin") > -1 ? "set-url" : "add";
 
+  await client.exec("git branch --unset-upstream", config, options);
+
   await client.exec(
     `git remote ${method} origin ${config.repository}`,
     config,
     options
   );
 
-  await client.exec("git fetch origin --prune", config, options);
+  const fetchResponse = await client.exec(
+    "git fetch origin --prune",
+    config,
+    options
+  );
 
-  await client.exec(`git checkout ${config.branch}`, config, options);
+  if (fetchResponse.success) {
+    await client.exec(`git checkout ${config.branch}`, config, options);
 
-  await client.exec("git reset --hard HEAD", config, options);
+    await client.exec("git reset --hard HEAD", config, options);
 
-  await client.exec(`git merge origin/${config.branch}`, config, options);
+    await client.exec(`git merge origin/${config.branch}`, config, options);
+  } else {
+    throw "git error";
+  }
 };
